Changing parameter agian again

+++ project-2_15-441/src/backend.c
@@ -326,7 +326,7 @@ void multi_send(cmu_socket_t *sock, uint8_t *data, int buf_len) {
     while (1) {
       int initial_left = left;
 
-      int timeout_length = (DEFAULT_TIMEOUT * 0.95) - (int)(current_timestamp_ms() - last_send_ts);
+      int timeout_length = (int)(DEFAULT_TIMEOUT * 0.95) - (int)(current_timestamp_ms() - last_send_ts);
       if (timeout_length <= 0) {
         break;
       }
